Modified the AT91SAM7S port and the ATmega168 port to use a slimmed down MS/TP state machine and datalink layer. Tested on AT91SAM7S-EK board.

+++ bacnet-stack/ports/at91sam7s/timer.h
@@ -35,7 +35,7 @@ extern "C" {
 #endif                          /* __cplusplus */
 
     void TimerInit(void);
-    uint16_t Timer_Silence(void);
+    int Timer_Silence(void);
     void Timer_Silence_Reset(void);
 
 #ifdef __cplusplus
